% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredplot.R
\name{fredplot}
\alias{fredplot}
\alias{fredplot.default}
\alias{fredplot.matrix}
\title{Functional residual density plot}
\usage{
fredplot(
  object,
  x,
  jitter = FALSE,
  subsample = NULL,
  resolution = 101,
  scale = c("uniform", "normal"),
  type = c("kde", "hex"),
  h = NULL,
  n = 100,
  plot = TRUE,
  color.palette = function(n) hcl.colors(n, "YlGnBu"),
  colorkey = FALSE,
  smooth = TRUE,
  smooth.col = "white",
  smooth.lwd = 1,
  smooth.lty = "solid",
  xlab = deparse1(substitute(x)),
  ylab = "Residual density",
  ...
)

\method{fredplot}{default}(
  object,
  x,
  jitter = FALSE,
  subsample = NULL,
  resolution = 101,
  scale = c("uniform", "normal"),
  type = c("kde", "hex"),
  h = NULL,
  n = 100,
  plot = TRUE,
  color.palette = function(n) hcl.colors(n, "YlGnBu"),
  colorkey = FALSE,
  smooth = TRUE,
  smooth.col = "white",
  smooth.lwd = 1,
  smooth.lty = "solid",
  xlab = deparse1(substitute(x)),
  ylab = "Residual density",
  ...
)

\method{fredplot}{matrix}(
  object,
  x,
  jitter = FALSE,
  subsample = NULL,
  resolution = 101,
  scale = c("uniform", "normal"),
  type = c("kde", "hex"),
  h = NULL,
  n = 100,
  plot = TRUE,
  color.palette = function(n) hcl.colors(n, "YlGnBu"),
  colorkey = FALSE,
  smooth = TRUE,
  smooth.col = "white",
  smooth.lwd = 1,
  smooth.lty = "solid",
  xlab = deparse1(substitute(x)),
  ylab = "Residual density",
  ...
)
}
\arguments{
\item{object}{Either an object for which the extraction of
\link[=fresiduals]{functional residuals} is meaningful (e.g., a \link[stats:glm]{glm}
object) or a two-column matrix giving the uniform lower/upper bounds for each
functional residual.}

\item{x}{Vector of predictor values to use for the x-axis. Ideally, \code{x}
should represent a continuous variable. Categorical variables should be
coerced to integer. Using \code{jitter()} can also be useful. For instance,
something like \code{x = jitter(as.integer(df$x))} could be useful. If \code{x} is
categorical in nature, it's also recommended to turn off LOESS smoothing by
setting \code{smooth = FALSE}.}

\item{jitter}{Logical indicating whether or not to add a small amount of
random noise to the predictor \code{x}. Default is \code{FALSE}. This can be useful
for visualizing categorical predictors or for avoiding issues with
\link[MASS:kde2d]{kde2d()} whenever \code{x} is discrete.}

\item{subsample}{Integer specifying the number of observations to draw at
random, which can help with compute time for very large datasets. Default
is \code{NULL} which corresponds to no subsampling.}

\item{resolution}{Integer specifying the resolution of the plot. Default
is 101.}

\item{scale}{Character string specifying which scale to use for plotting.
Default is \code{"uniform"}.}

\item{type}{Character string specifying the type of plot to construct.
Current options include \code{"hex"} (the default) for a hexagonal heatmap
of 2D bin counts, or \code{"kde"} for 2D kernel density estimation.}

\item{h}{Vector of bandwidths passed to \link[MASS:kde2d]{kde2d()}. Default is
\code{NULL} which will result in using the normal reference bandwidth. See
\link[MASS:kde2d]{?MASS::kde2d} for details.}

\item{n}{Integer specifying the number of grid points in each direction for
two-dimensional kernel density estimation; passed to \link[MASS:kde2d]{kde2d()}
whenever \code{type = "kde"}. Default is 100.}

\item{plot}{Logical indicating whether to return a data frame that can be
plotted manually (\code{FALSE}) or plot the results automatically (\code{TRUE}).
Default is \code{TRUE}. Setting to \code{FALSE} is useful for debugging or advanced
usage.}

\item{color.palette}{A color palette function to be used to assign colors in
the plot.}

\item{colorkey}{Logical indicating whether or not a legend should be drawn.
Default is \code{FALSE}.}

\item{smooth}{Logical indiacting whether or not to include a LOESS smoother.
Default is \code{TRUE},}

\item{smooth.col}{Integer or character string specifying the line color to
use for the optional LOESS smoother; default is \code{"white"}.}

\item{smooth.lwd}{Integer specifying the line width to use for the optional
LOESS smoother; default is 1.}

\item{smooth.lty}{Integer or character string specifying the line type to
use for the optional LOESS smoother; default is 1 (equivalent to \code{"solid"}).}

\item{xlab, ylab}{Character strings specifying the labels to use for the x-
and y-axes, respectively.}

\item{...}{Additional optional arguments passed to
\link[lattice:levelplot]{lattice::levelplot()} (\code{type = "kde"}) or
\link[hexbin:hexbinplot]{hexbin::hexbinplot()} (\code{type = "hex"}).}
}
\value{
A \code{trellis} object (see \link[lattice:Lattice]{lattice::lattice}).
}
\description{
Generates a functional residual-vs-predictor plot, visualizing the density of
the functional residuals as a heatmap rendered on either the uniform or
normal scale.
}
\examples{
# Generate data from a logistic regression model with quadratic form
set.seed(1217)
n <- 1000
x <- rnorm(n)
z <- 1 - 2*x + 3*x^2 + rlogis(n)
y <- ifelse(z > 0, 1, 0)

# Fit models with/without quadratic term
bad <- glm(y ~ x, family = binomial)  # wrong
good <- glm(y ~ x + I(x^2), family = binomial)  # right

# Functional REsidual Density plot for each model
gridExtra::grid.arrange(
  fredplot(bad, x = x, type = "hex", aspect = 1),
  fredplot(good, x = x, type = "hex", aspect = 1),
  nrow = 1
)
}
\references{
Liu, D., Lin, Z., & Zhang, H. (2025). A unified framework for residual
diagnostics in generalized linear models and beyond. \emph{Journal of the American
Statistical Association}, 1–29.
\doi{10.1080/01621459.2025.2504037}
}
